changed doubles to floats; added RFC template

+++ CPP_flight_computer_program/include/sensor_ms5611.h
@@ -10,6 +10,6 @@ float ms5611_ground_base_pressure = FLO_DEF;
 bool init_MS5611();
 float get_ms5611_temp();
 float get_ms5611_press();
-double get_ms5611_altitude(double pressure, double seaLevelPressure);
+float get_ms5611_altitude(float pressure, float seaLevelPressure);
 
 #endif // SENSOR_MS5611_H
